Refactor, add better error handling
