+++ tests/integration/snowflake.spec.ts
@@ -9,7 +9,7 @@ import { dropAllTables, getTableRows, keyBy } from "df/tests/integration/utils";
 suite("@dataform/integration/snowflake", ({ tearDown }) => {
   const credentials = dfapi.credentials.read("snowflake", "test_credentials/snowflake.json");
   const dbadapter = dbadapters.create(credentials, "snowflake");
-  tearDown(() => dbadapter.close());
+  tearDown("close adapter", () => dbadapter.close());
 
   test("run", { timeout: 60000 }, async () => {
     const compiledGraph = await dfapi.compile({
